Diffs actually work now

+++ core/src/animated/spferical/netrogue/networking/NetworkObject.java
@@ -19,7 +19,7 @@ public abstract class NetworkObject implements Serializable, Cloneable {
 	 * 
 	 */
 	private static final long serialVersionUID = -4277851132396689715L;
-	public int lastUpdate;
+	public long lastUpdate;
 	public long ID;
 	public NetworkObject parent;
